requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pandas==2.1.4
fpdf2==2.7.6
openpyxl==3.1.2

